+++ app/src/main/java/com/example/emptyactivity/home/Account.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity
+package com.example.emptyactivity.home
 
 // Import necessary Compose components and libraries
 import androidx.compose.foundation.layout.Arrangement
@@ -23,7 +23,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import java.lang.NumberFormatException
 
 @Composable
 fun StatefulAccount(modifier: Modifier = Modifier) {
